requirements.txt
pyspark==3.4.1
delta-spark==2.4.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
mlflow==2.9.2
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
py4j==0.10.9.7 